+++ SendBirdSDK.framework/Headers/SBDOpenChannelListQuery.h
@@ -42,12 +42,25 @@ DEPRECATED_ATTRIBUTE;
 DEPRECATED_ATTRIBUTE;
 
 /**
- *  Sets the custom type filter.
+ * Sets the channel URL filter.
  *
- *  @deprecated in 3.0.64. Use `setCustomTypeFilter:` instead.
+ *  @see combined with `setChannelUrlFilter:` method.
  */
-@property (strong, nonatomic, nullable) NSString *customTypeFilter
-DEPRECATED_ATTRIBUTE;
+@property (copy, nonatomic, nullable) NSString *channelUrlFilter;
+
+/**
+ * Sets the channel name filter.
+ *
+ *  @see combined with `setChannelNameFilter:` method.
+ */
+@property (copy, nonatomic, nullable) NSString *channelNameFilter;
+
+/**
+ * Sets the custom type filter.
+ *
+ *  @see combined with `setCustomTypeFilter:` method.
+ */
+@property (copy, nonatomic, nullable) NSString *customTypeFilter;
 
 /**
  *  Shows if the query is loading.
@@ -55,30 +68,6 @@ DEPRECATED_ATTRIBUTE;
  *  @return Returns YES if the query is loading, otherwise returns NO.
  */
 - (BOOL)isLoading;
-
-
-/**
- Sets the channel URL filter.
-
- @param channelUrl The channel url to search.
- */
-- (void)setChannelUrlFilter:(NSString * _Nullable)channelUrl;
-
-
-/**
- Sets the channel name filter.
-
- @param channelName The channel name to search.
- */
-- (void)setChannelNameFilter:(NSString * _Nullable)channelName;
-
-
-/**
- Sets the custom type filter.
-
- @param customType The custom type to search.
- */
-- (void)setCustomTypeFilter:(NSString * _Nullable)customType;
 
 /**
  *  Gets the list of channels. If this method is repeatedly called, it will retrieve the following pages of the channel list.